requirements.txt
gradio>=3.0
tensorflow>=2.10.0
numpy
Pillow